# docs/api/openapi.yaml
openapi: 3.1.0
info:
  title: DeenMate API
  description: |
    Comprehensive Islamic application backend providing access to Quran, Hadith, Prayer Times, 
    Zakat calculations, and Islamic audio content. This API is designed to serve Islamic 
    applications, websites, and mobile apps with accurate, reliable, and well-structured Islamic data.
  
    ## Features
    - **Quran Module**: Complete Quran text, translations, and recitations
    - **Prayer Module**: Accurate prayer times, qibla direction, and calculation methods
    - **Hadith Module**: Authentic hadith collections and search functionality
    - **Zakat Module**: Zakat calculations and current market values
    - **Audio Module**: High-quality Islamic audio content and metadata
  
    ## Authentication
    Currently operates in read-only mode without authentication. Future versions will include 
    JWT-based authentication for user-specific features.
  
    ## Rate Limiting
    - Public Endpoints: 100 requests per minute per IP
    - Authenticated Endpoints: 1000 requests per minute per user
    - Premium Tier: 5000 requests per minute per user
  
    ## Support
    - Documentation: https://docs.deenmate.com
    - API Status: https://status.deenmate.com
    - Support Email: api-support@deenmate.com
    - Developer Discord: https://discord.gg/deenmate
  
  version: 1.0.0
  contact:
    name: DeenMate API Support
    email: api-support@deenmate.com
    url: https://docs.deenmate.com
  license:
...
servers:
...
paths:
...
  /prayer/times/{date}:
    get:
      tags:
        - Prayer
      summary: Get Daily Prayer Times
      description: Upstream-compatible with aladhan.com/v1/timings/{date}
      parameters:
        - name: date
          in: path
          required: true
          description: Date in YYYY-MM-DD format
          schema:
            type: string
            format: date
        - name: lat
          in: query
          required: true
          description: Latitude
          schema:
            type: number
            format: float
        - name: lng
          in: query
          required: true
          description: Longitude
          schema:
            type: number
            format: float
        - name: method
          in: query
          schema:
            type: integer
            default: 1
        - name: school
          in: query
          schema:
            type: integer
            default: 0
        - name: latitudeAdjustmentMethod
          in: query
          schema:
            type: integer
            enum: [0, 1, 2]
        - name: tune
          in: query
          schema:
            type: string
            description: Comma-separated 7 integer minute offsets
        - name: timezonestring
          in: query
          schema:
            type: string
            description: IANA timezone (e.g., Asia/Dhaka)
      responses:
        '200':
          description: Daily prayer times
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PrayerTimesResponse'
        '400':
          description: Invalid parameters
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /prayer/times/{locKey}/monthly/{year}/{month}:
    get:
      tags:
        - Prayer
      summary: Get Monthly Prayer Calendar
      description: Upstream-compatible with aladhan.com/v1/calendar/{year}/{month}
      parameters:
        - name: locKey
          in: path
          required: true
          description: DeenMate location key (hash of lat/lng)
          schema:
            type: string
        - name: year
          in: path
          required: true
          schema:
            type: integer
        - name: month
          in: path
          required: true
          schema:
            type: integer
            minimum: 1
            maximum: 12
        - name: method
          in: query
          schema:
            type: integer
            default: 1
        - name: school
          in: query
          schema:
            type: integer
            default: 0
        - name: lat
          in: query
          description: Latitude (for upstream fallback only)
          schema:
            type: number
            format: float
        - name: lng
          in: query
          description: Longitude (for upstream fallback only)
          schema:
            type: number
            format: float
        - name: latitudeAdjustmentMethod
          in: query
          schema:
            type: integer
            enum: [0, 1, 2]
        - name: tune
          in: query
          schema:
            type: string
            description: Comma-separated 7 integer minute offsets
        - name: timezonestring
          in: query
          schema:
            type: string
            description: IANA timezone (e.g., Asia/Dhaka)
      responses:
        '200':
          description: Monthly prayer calendar
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/PrayerTimes'
  /prayer/qibla:
    get:
      tags:
        - Prayer
      summary: Get Qibla Direction
      description: Calculate qibla direction from a location
      parameters:
        - name: latitude
          in: query
          required: true
          description: Latitude coordinate
          schema:
            type: number
            format: float
        - name: longitude
          in: query
          required: true
          description: Longitude coordinate
          schema:
            type: number
            format: float
      responses:
        '200':
          description: Qibla direction information
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/QiblaResponse'
...
components:
  schemas:
...
    BaseResponse:
      type: object
      properties:
        success:
          type: boolean
          description: Whether the request was successful
        data:
          description: Response data
        meta:
          description: Metadata about the response
      required:
        - success
    ErrorResponse:
      allOf:
        - $ref: '#/components/schemas/BaseResponse'
        - type: object
          properties:
            success:
              const: false
            error:
              type: string
              description: Error message
            code:
              type: string
              description: Error code
            details:
              description: Additional error details
          required:
            - error
...
    PrayerTimes:
      type: object
      properties:
        fajr:
          type: string
          description: Fajr prayer time (HH:MM)
        sunrise:
          type: string
          description: Sunrise time (HH:MM)
        dhuhr:
          type: string
          description: Dhuhr prayer time (HH:MM)
        asr:
          type: string
          description: Asr prayer time (HH:MM)
        maghrib:
          type: string
          description: Maghrib prayer time (HH:MM)
        isha:
          type: string
          description: Isha prayer time (HH:MM)
        method:
          type: string
          description: Calculation method used
    PrayerTimesResponse:
      allOf:
        - $ref: '#/components/schemas/BaseResponse'
        - type: object
          properties:
            success:
              const: true
            data:
              type: object
              properties:
                prayerTimes:
                  $ref: '#/components/schemas/PrayerTimes'
                location:
                  type: object
                  properties:
                    latitude:
                      type: number
                    longitude:
                      type: number
                method:
                  type: integer
                  description: Calculation method ID
                date:
                  type: string
                  format: date
            meta:
              type: object
              properties:
                cacheTtl:
                  type: string
                  description: Cache TTL information
                calculationMethod:
                  type: string
                  description: Calculation method name

    # Zakat Schemas
...
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT token for authenticated endpoints

# Global security (for future use)
# security:
#   - BearerAuth: []